linear GP transport over

+++ 2D_transport/init.h
@@ -1,5 +1,11 @@
 #include "parameters.h"
-#include "2D_bij.h"
 
-
-__global__ void initialize(float f[], float x[], float y[], const float dx, const float dy);
+/*
+* initialize : initialize the mesh and the solution
+* f[]  : pointer to the solution
+* x[]  : pointer to x mesh
+* y[]  : pointer to y mesh
+* dx   : x step
+* dy   : y step
+*/
+__global__ void initialize(double f[], double x[], double y[], const double dx, const double dy);
